requirements/common.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file=requirements/common.txt requirements/common.in
#
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
click==7.1.2
    # via flask
flask-limiter==1.4
    # via -r requirements/common.in
flask==1.1.2
    # via
    #   -r requirements/common.in
    #   flask-limiter
gunicorn==20.0.4
    # via -r requirements/common.in
idna==2.10
    # via requests
itsdangerous==1.1.0
    # via flask
jinja2==2.11.3
    # via flask
limits==1.5.1
    # via flask-limiter
markupsafe==1.1.1
    # via jinja2
marshmallow==3.10.0
    # via webargs
pillow==8.1.2
    # via -r requirements/common.in
requests==2.25.1
    # via -r requirements/common.in
six==1.15.0
    # via
    #   flask-limiter
    #   limits
urllib3==1.26.4
    # via requests
webargs==7.0.1
    # via -r requirements/common.in
werkzeug==1.0.1
    # via flask

# The following packages are considered to be unsafe in a requirements file:
# setuptools
